feat(footer): implement complete footer component structure with social links and branding
