+++ apps/frontend/src/components/Footer/Footer.test.tsx
@@ -1,6 +1,6 @@
 import "@testing-library/jest-dom"
 import { render, screen } from "@/test-config/test-utils"
-import { Footer } from "./Footer"
+import { Footer } from "."
 
 describe("<Footer />", () => {
   it("should have correct displayName", () => {
